Fixed a bug that I had created... There was a circular reference in the import statements with the Workbook import due to circular dependencies. I needed to move an refactor the ExcelError from main to tools and the other way refactor excel2df from tools to main
